cleanup and fixed

+++ .travis/deploy.sh
@@ -2,6 +2,12 @@
 
 set -e
 
+# only do deployment, when travis detects a new tag
+if [ ! -z "$TRAVIS_TAG" ]
+then
+    echo "on a tag -> set pom.xml <version> to $TRAVIS_TAG"
+    mvn --settings .travis/settings.xml org.codehaus.mojo:versions-maven-plugin:2.3:set -DnewVersion=$TRAVIS_TAG -Prelease
+
     if [ ! -z "$TRAVIS" -a -f "$HOME/.gnupg" ]; then
         shred -v ~/.gnupg/*
         rm -rf ~/.gnupg
@@ -9,20 +15,12 @@ set -e
 
     source .travis/gpg.sh
 
+    mvn clean deploy --settings .travis/settings.xml -DskipTests=true -B -U -Prelease
+
     if [ ! -z "$TRAVIS" ]; then
         shred -v ~/.gnupg/*
         rm -rf ~/.gnupg
     fi
-# only do deployment, when travis detects a new tag
-if [ ! -z "$TRAVIS_TAG" ]
-then
-    echo "on a tag -> set pom.xml <version> to $TRAVIS_TAG"
-    mvn --settings .travis/settings.xml org.codehaus.mojo:versions-maven-plugin:2.3:set -DnewVersion=$TRAVIS_TAG -Prelease
-
-
-    mvn clean deploy --settings .travis/settings.xml -DskipTests=true -B -U -Prelease
-
-
 else
     echo "not on a tag -> keep snapshot version in pom.xml"
 fi
